implement delete functionality and adding and working on update one

+++ src/pages/MyList/MyList.jsx
@@ -1,6 +1,7 @@
 import { useContext, useState } from "react";
-import { useLoaderData } from "react-router-dom";
+import { Link, useLoaderData } from "react-router-dom";
 import { AuthContex } from "../../Providers/AuthProvider";
+import Swal from "sweetalert2";
 
 const MyList = () => {
   const allMyListData = useLoaderData();
@@ -20,20 +21,41 @@ const MyList = () => {
   console.log(remaining);
 
   const handleDelete = (id) => {
-    //console.log(id);
-    fetch(`http://localhost:5000/touristSpot/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          console.log('Deleted successfully');
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        //console.log(id);
+        fetch(`http://localhost:5000/touristSpot/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              console.log("Deleted successfully");
+              Swal.fire({
+                position: "top-end",
+                icon: "success",
+                title: "Delete Successful",
+                showConfirmButton: false,
+                timer: 1500,
+              });
 
-          //remove the users from the ui
-          const remainingUsers = loadeduser.filter(user => user._id !== id);
-          setLoadedUser(remainingUsers)
-        }
-      });
+              //remove the users from the ui
+              const remainingUsers = loadeduser.filter(
+                (user) => user._id !== id
+              );
+              setLoadedUser(remainingUsers);
+            }
+          });
+      }
+    });
   };
 
   // const {
@@ -93,9 +115,13 @@ const MyList = () => {
                   </td>
 
                   <td className="p-3 mr-6 text-right">
+                  {/* to={`/tourSpot/${_id}`} */}
+
+                    <Link to={`/updateTourSpot/${singleSpot._id}`}>
                     <button className="px-3 py-2 font-semibold rounded-md bg-violet-600 text-gray-50 mb-2">
                       Update
                     </button>
+                    </Link>
                     <button
                       onClick={() => handleDelete(singleSpot._id)}
                       className="px-3 py-2 font-semibold rounded-md bg-red-800 text-gray-50"
